Do not overwrite src/test/install/setup.properties if it exist

+++ src/test/scripts/prepare_test.py
@@ -41,12 +41,13 @@ finally:
 for f in glob.glob("src/test/install/*.war"):
     os.remove(f)
 
-with open("src/test/install/setup.properties", "wt") as f:
-    print("secure         = true", file=f)
-    print("container      = Glassfish", file=f)
-    print("home           = %s" % containerHome, file=f)
-    print("port           = 4848", file=f)
-    print("libraries      = ids.storage_test*.jar", file=f)
+if not os.path.exists("src/test/install/setup.properties"):
+    with open("src/test/install/setup.properties", "wt") as f:
+        print("secure         = true", file=f)
+        print("container      = Glassfish", file=f)
+        print("home           = %s" % containerHome, file=f)
+        print("port           = 4848", file=f)
+        print("libraries      = ids.storage_test*.jar", file=f)
 
 with open("src/test/install/run.properties.example", "wt") as f:
     pass
